requirements.txt
Pillow
pytesseract
pyautogui
